+++ src/mesh/amr_loadbalance.cpp
@@ -696,6 +696,8 @@ void Mesh::RedistributeAndRefineMeshBlocks(ParameterInput *pin, ApplicationInput
   }
   Initialize(2, pin, app_in);
 
+  BuildMeshBlockPacks();
+
   ResetLoadBalanceVariables();
 
   return;