+++ src/utils/partition_stl_containers.hpp
@@ -42,17 +42,17 @@ int IntCeil(int x, int y) {
 // Takes container of elements and fills partitions
 // of size N with pointers to elements.
 // Assumes Container_t has STL-style iterators defined
-template <typename Container_t, typename T>
-void ToSizeN(Container_t &container, const int N, Partition_t<T> &partitions) {
+template <typename T, template <class...> typename Container_t, class... extra>
+auto ToSizeN(Container_t<T, extra...> &container, const int N) {
   using std::to_string;
   using namespace partition_impl;
 
-  PARTHENON_REQUIRE_THROWS(N > 0, "You must have at least 1 partition");
+  PARTHENON_REQUIRE_THROWS(N > 0, "Your partition must be at least size 1");
 
   int nelements = container.size();
   int npartitions = IntCeil(nelements, N);
 
-  partitions.resize(npartitions);
+  Partition_t<T> partitions(npartitions);
   for (auto &p : partitions) {
     p.reserve(N);
     p.clear();
@@ -67,17 +67,18 @@ void ToSizeN(Container_t &container, const int N, Partition_t<T> &partitions) {
       b = 0;
     }
   }
+  return partitions;
 }
 
 // Takes container of elements and fills N partitions
 // with pointers to elements.
 // Assumes Container_t has STL-style iterators defined
-template <typename Container_t, typename T>
-void ToNPartitions(Container_t &container, const int N, Partition_t<T> &partitions) {
+template <typename T, template <class...> typename Container_t, class... extra>
+auto ToNPartitions(Container_t<T, extra...> &container, const int N) {
   using namespace partition_impl;
   int nelements = container.size();
   int partition_size = IntCeil(nelements, N);
-  ToSizeN(container, partition_size, partitions);
+  return ToSizeN(container, partition_size);
 }
 } // namespace partition
 } // namespace parthenon